Extract hook for ref option querying
